chore: bulk export fix, admin part js refactoring
