+++ src/main/resources/webapp/pdf-exporter-admin/pages/cover-page.jsp
@@ -12,6 +12,7 @@
     <link rel="stylesheet" href="../ui/generic/css/common.css?bundle=<%= bundleTimestamp %>">
     <link rel="stylesheet" href="../ui/generic/css/custom-select.css?bundle=<%= bundleTimestamp %>">
     <link rel="stylesheet" href="../ui/generic/css/configurations.css?bundle=<%= bundleTimestamp %>">
+    <script type="module" src="../js/modules/cover-page.js?bundle=<%= bundleTimestamp %>"></script>
     <style type="text/css">
         html {
             height: 100%;
@@ -85,11 +86,7 @@
     <input id="bundle-timestamp" type="hidden" value="<%= ch.sbb.polarion.extension.generic.util.VersionUtils.getVersion().getBundleBuildTimestamp() %>"/>
 </div>
 
-<jsp:include page='/common/jsp/buttons.jsp'>
-    <jsp:param name="saveFunction" value="saveCoverPage()"/>
-    <jsp:param name="cancelFunction" value="SbbCommon.cancelEdit()"/>
-    <jsp:param name="defaultFunction" value="revertToDefault()"/>
-</jsp:include>
+<jsp:include page='/common/jsp/buttons.jsp'/>
 
 <div class="standard-admin-page hide-on-edit-configuration" id="templates-pane" style="display: none">
     <h2 class="align-left">Predefined Templates</h2>
@@ -99,7 +96,7 @@
         <label for="templates-select" id="templates-label">Predefined Templates:</label>
         <div id="templates-select"></div>
         <div class="action-buttons">
-            <button class="toolbar-button" onclick="Templates.persistSelected()">
+            <button id="persist-selected-template" class="toolbar-button">
                 <img class="button-image" src="/polarion/ria/images/actions/save.gif?bundle=<%= bundleTimestamp %>" alt="">Persist
             </button>
         </div>
@@ -131,10 +128,5 @@
         <jsp:include page="../pages/placeholders.jsp"/>
     </div>
 </div>
-
-<script type="text/javascript" src="../ui/generic/js/common.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../ui/generic/js/custom-select.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../ui/generic/js/configurations.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../js/cover-page.js?bundle=<%= bundleTimestamp %>"></script>
 </body>
 </html>
