+++ src/main/resources/webapp/pdf-exporter-admin/pages/style-packages.jsp
@@ -14,6 +14,7 @@
     <link rel="stylesheet" href="../ui/generic/css/common.css?bundle=<%= bundleTimestamp %>">
     <link rel="stylesheet" href="../ui/generic/css/custom-select.css?bundle=<%= bundleTimestamp %>">
     <link rel="stylesheet" href="../ui/generic/css/configurations.css?bundle=<%= bundleTimestamp %>">
+    <script type="module" src="../js/modules/style-packages.js?bundle=<%= bundleTimestamp %>"></script>
     <style type="text/css">
         html {
             height: 100%;
@@ -95,7 +96,7 @@
                 <div class='input-group flex-centered'>
                     <label for='style-package-weight'>Weight:</label>
                     <div class='more-info' title="A float number from 0.0 to 100, which will determine the position of current style package in the resulting style packages list. The higher the number, the higher its position will be."></div>
-                    <input id="style-package-weight" style="margin-left: 59px" type="number" min="1" max="100" step="0.1" onblur="StylePackageUtils.adjustWeight(this)">
+                    <input id="style-package-weight" style="margin-left: 59px" type="number" min="1" max="100" step="0.1">
                 </div>
             </div>
 
@@ -290,16 +291,7 @@
     <input id="bundle-timestamp" type="hidden" value="<%= ch.sbb.polarion.extension.generic.util.VersionUtils.getVersion().getBundleBuildTimestamp() %>"/>
 </div>
 
-<jsp:include page='/common/jsp/buttons.jsp'>
-    <jsp:param name="saveFunction" value="saveStylePackage()"/>
-    <jsp:param name="cancelFunction" value="SbbCommon.cancelEdit()"/>
-    <jsp:param name="defaultFunction" value="revertToDefault()"/>
-</jsp:include>
+<jsp:include page='/common/jsp/buttons.jsp'/>
 
-<script type="text/javascript" src="../ui/generic/js/common.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../ui/generic/js/custom-select.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../ui/generic/js/configurations.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../js/style-packages.js?bundle=<%= bundleTimestamp %>"></script>
-<script type="text/javascript" src="../js/style-package-utils.js?bundle=<%= bundleTimestamp %>"></script>
 </body>
 </html>
